+++ CHANGELOG.md
@@ -5,6 +5,7 @@ New functionality added in a backwards-compatible manner:
 
 * [OLMIS-1693:](https://openlmis.atlassian.net/browse/OLMIS-2704) Added openlmis-invalid and error message documentation.
 * [OLMIS-249:](https://openlmis.atlassian.net/browse/OLMIS-249) Datepicker element now allows translating day and month names.
+* [OLMIS-2817:](https://openlmis.atlassian.net/browse/OLMIS-2817) Added new file input directive.
 
 5.1.0 / 2017-06-22
 ========================